Move back script to view, and move vue component from all sales_order to app.
