+++ resources/assets/js/app.js
@@ -5,9 +5,9 @@
  * building robust, powerful web applications using Vue and Laravel.
  */
 
-require('./bootstrap');
+// require('./bootstrap');
 
-window.Vue = require('vue');
+// window.Vue = require('vue');
 
 /**
  * Next, we will create a fresh Vue application instance and attach it to
@@ -15,9 +15,13 @@ window.Vue = require('vue');
  * or customize the JavaScript scaffolding to fit your unique needs.
  */
 
-Vue.component('example', require('./components/Example.vue'));
+Vue.use(VeeValidate, { locale: $('html').attr('lang') });
 
+Vue.component('example', require('./components/Example.vue'));
 Vue.component('vue-numeric', require('./components/vue-numeric.vue'));
+Vue.component('vue-icheck', require('./components/vue-icheck.vue'));
+Vue.component('vue-datetimepicker', require('./components/vue-datetimepicker.vue'));
+Vue.component('select2_customer', require('./components/select2_customer.vue'));
 
 const app = new Vue({
     el: '#app'
