+++ resources/views/layouts/adminlte/master.blade.php
@@ -59,8 +59,8 @@
 
         <input type="hidden" id="secapi" value="{{ Auth::user()->api_token }}">
 
-        <!--<script type="application/javascript" src="{{ asset('adminlte/js/app.js') }}"></script>-->
         <script type="application/javascript" src="{{ asset('adminlte/js/adminlte.js') }}"></script>
+        <script type="application/javascript" src="{{ asset('adminlte/js/app.js') }}"></script>
 
         <script type="application/javascript">
             $(document).ready(function () {
