+++ resources/views/sales_order/payment/broughtforward_payment.blade.php
@@ -89,8 +89,6 @@
 
 @section('custom_js')
     <script type="application/javascript">
-        Vue.use(VeeValidate, { locale: '{!! LaravelLocalization::getCurrentLocale() !!}' });
-
         var vm = new Vue({
             el: '#soPaymentVue',
             data: {
